requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
Django==5.1.4
djangorestframework==3.16.1
gunicorn==23.0.0
h11==0.16.0
idna==3.10
jellyfish==1.2.0
lib50==3.1.1
markdown2==2.5.3
outcome==1.3.0.post0
packaging==25.0
pexpect==4.9.0
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
selenium==4.35.0
setuptools==80.9.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
submit50==3.2.1
termcolor==1.1.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2024.2
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
